refactor: hashのドメインエンティティのディレクトリを変更
