+++ pkg/domain/hash/IHashAggregateRoot.hpp
